UPD: https://chromium.googlesource.com/chromium/src/base/+/944910a3c30b76db99a3978f59098e8a33953617

+++ base/android/java/src/org/chromium/base/FileUtils.java
@@ -9,9 +9,11 @@ import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 import android.net.Uri;
 import android.os.ParcelFileDescriptor;
-import android.support.annotation.Nullable;
 
-import java.io.BufferedOutputStream;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+
+import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.FileDescriptor;
 import java.io.FileOutputStream;
@@ -27,37 +29,60 @@ import java.util.Locale;
 public class FileUtils {
     private static final String TAG = "FileUtils";
 
+    public static Function<String, Boolean> DELETE_ALL = filepath -> true;
+
     /**
      * Delete the given File and (if it's a directory) everything within it.
+     * @param currentFile The file or directory to delete. Does not need to exist.
+     * @param canDelete the {@link Function} function used to check if the file can be deleted.
+     * @return True if the files are deleted, or files reserved by |canDelete|, false if failed to
+     *         delete files.
+     * @note Caveat: Return values from recursive deletes are ignored.
+     * @note Caveat: |canDelete| is not robust; see https://crbug.com/1066733.
      */
-    public static void recursivelyDeleteFile(File currentFile) {
-        ThreadUtils.assertOnBackgroundThread();
+    public static boolean recursivelyDeleteFile(
+            File currentFile, Function<String, Boolean> canDelete) {
+        if (!currentFile.exists()) {
+            // This file could be a broken symlink, so try to delete. If we don't delete a broken
+            // symlink, the directory containing it cannot be deleted.
+            currentFile.delete();
+            return true;
+        }
+        if (canDelete != null && !canDelete.apply(currentFile.getPath())) {
+            return true;
+        }
+
         if (currentFile.isDirectory()) {
             File[] files = currentFile.listFiles();
             if (files != null) {
                 for (File file : files) {
-                    recursivelyDeleteFile(file);
+                    recursivelyDeleteFile(file, canDelete);
                 }
             }
         }
 
-        if (!currentFile.delete()) Log.e(TAG, "Failed to delete: " + currentFile);
+        boolean ret = currentFile.delete();
+        if (!ret) {
+            Log.e(TAG, "Failed to delete: %s", currentFile);
+        }
+        return ret;
     }
 
     /**
      * Delete the given files or directories by calling {@link #recursivelyDeleteFile(File)}. This
      * supports deletion of content URIs.
      * @param filePaths The file paths or content URIs to delete.
+     * @param canDelete the {@link Function} function used to check if the file can be deleted.
      */
-    public static void batchDeleteFiles(List<String> filePaths) {
-        ThreadUtils.assertOnBackgroundThread();
-
+    public static void batchDeleteFiles(
+            List<String> filePaths, Function<String, Boolean> canDelete) {
         for (String filePath : filePaths) {
+            if (canDelete != null && !canDelete.apply(filePath)) continue;
             if (ContentUriUtils.isContentUri(filePath)) {
                 ContentUriUtils.delete(filePath);
             } else {
                 File file = new File(filePath);
-                if (file.exists()) recursivelyDeleteFile(file);
+                if (file.exists()) recursivelyDeleteFile(file, canDelete);
             }
         }
     }
@@ -66,62 +91,55 @@ public class FileUtils {
      * Extracts an asset from the app's APK to a file.
      * @param context
      * @param assetName Name of the asset to extract.
-     * @param dest File to extract the asset to.
+     * @param outFile File to extract the asset to.
      * @return true on success.
      */
-    public static boolean extractAsset(Context context, String assetName, File dest) {
-        InputStream inputStream = null;
-        OutputStream outputStream = null;
-        try {
-            inputStream = context.getAssets().open(assetName);
-            outputStream = new BufferedOutputStream(new FileOutputStream(dest));
-            byte[] buffer = new byte[8192];
-            int c;
-            while ((c = inputStream.read(buffer)) != -1) {
-                outputStream.write(buffer, 0, c);
-            }
-            inputStream.close();
-            outputStream.close();
+    public static boolean extractAsset(Context context, String assetName, File outFile) {
+        try (InputStream inputStream = context.getAssets().open(assetName)) {
+            copyStreamToFile(inputStream, outFile);
             return true;
         } catch (IOException e) {
-            if (inputStream != null) {
-                try {
-                    inputStream.close();
-                } catch (IOException ex) {
-                }
-            }
-            if (outputStream != null) {
-                try {
-                    outputStream.close();
-                } catch (IOException ex) {
-                }
-            }
+            return false;
         }
-        return false;
+    }
+
+    /**
+     * Performs a simple copy of inputStream to outputStream.
+     */
+    public static void copyStream(InputStream inputStream, OutputStream outputStream)
+            throws IOException {
+        byte[] buffer = new byte[8192];
+        int amountRead;
+        while ((amountRead = inputStream.read(buffer)) != -1) {
+            outputStream.write(buffer, 0, amountRead);
+        }
     }
 
     /**
      * Atomically copies the data from an input stream into an output file.
      * @param is Input file stream to read data from.
      * @param outFile Output file path.
-     * @param buffer Caller-provided buffer. Provided to avoid allocating the same
-     *         buffer on each call when copying several files in sequence.
      * @throws IOException in case of I/O error.
      */
-    public static void copyFileStreamAtomicWithBuffer(InputStream is, File outFile, byte[] buffer)
-            throws IOException {
+    public static void copyStreamToFile(InputStream is, File outFile) throws IOException {
         File tmpOutputFile = new File(outFile.getPath() + ".tmp");
         try (OutputStream os = new FileOutputStream(tmpOutputFile)) {
             Log.i(TAG, "Writing to %s", outFile);
-
-            int count = 0;
-            while ((count = is.read(buffer, 0, buffer.length)) != -1) {
-                os.write(buffer, 0, count);
-            }
+            copyStream(is, os);
         }
         if (!tmpOutputFile.renameTo(outFile)) {
             throw new IOException();
         }
+    }
+
+    /**
+     * Reads inputStream into a byte array.
+     */
+    @NonNull
+    public static byte[] readStream(InputStream inputStream) throws IOException {
+        ByteArrayOutputStream data = new ByteArrayOutputStream();
+        FileUtils.copyStream(inputStream, data);
+        return data.toByteArray();
     }
 
     /**
@@ -149,13 +167,14 @@ public class FileUtils {
 
     /**
      * Returns the file extension, or an empty string if none.
-     * @param file Name of the file, with or without the full path.
+     * @param file Name of the file, with or without the full path (Unix style).
      * @return empty string if no extension, extension otherwise.
      */
     public static String getExtension(String file) {
-        int index = file.lastIndexOf('.');
-        if (index == -1) return "";
-        return file.substring(index + 1).toLowerCase(Locale.US);
+        int lastSep = file.lastIndexOf('/');
+        int lastDot = file.lastIndexOf('.');
+        if (lastSep >= lastDot) return ""; // Subsumes |lastDot == -1|.
+        return file.substring(lastDot + 1).toLowerCase(Locale.US);
     }
 
     /** Queries and decodes bitmap from content provider. */
